Kirk's recommended changes
exercise9
Make docstring indentation consistent with the ''' indentation (in this case for characters). It makes it easier to read.
There is pretty much always a better solution than using 'eval'. Here are a couple of alternate solutions (that are better):
getattr(pynet, 'rtr1')           # This would get the 'rtr1' dictionary from the pynet module

'rtr1' could also be a variable that is contains a string (i.e. 'rtr1', 'rtr2', 'srx1', but in a variable)

You could also stick the 'hostname' in the dictionary in pynet.py and then pop() it off the dictionary.

hostname = a_device.pop('hostname')

This would affect other code that was using the pynet.py module, however.